+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-01T14:13:53.072Z",
+    "2026-03-02T09:56:02.262Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {