+++ packages/orq-rc/RELEASES.md
@@ -7379,3 +7379,13 @@ Based on:
 - [typescript v4.5.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.17 - packages/orq-rc
+
+## 2026-03-02 09:59:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.736.3 (2.845.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.18 - packages/orq-rc